refactor: 重构文件结构并优化代码组织
- 将核心模块和工具类移至`core`目录下，提升代码可维护性
- 重构`section`基类，增加类型提示和文档注释
- 优化`BinaryReader`类，增强错误处理和资源管理
- 更新文件路径引用，确保模块间依赖清晰
