+++ src/league_tools/cli/__init__.py
@@ -3,10 +3,6 @@
 # @Email   : Virace@aliyun.com
 # @Site    : x-item.com
 # @Software: Pycharm
-# @Create  : 2024/3/24 1:12
-# @Update  : 2024/3/24 1:12
+# @Create  : 2025/4/26 3:11
+# @Update  : 2025/4/26 3:11
 # @Detail  : 
-
-from .reader import BinaryReader
-
-__all__ = BinaryReader
