+++ src/league_tools/core/__init__.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
-# @Time    : 2021/2/28 14:43
 # @Author  : Virace
 # @Email   : Virace@aliyun.com
 # @Site    : x-item.com
 # @Software: PyCharm
-# @Detail  :
+# @Create  : 2021/2/28 14:43
+# @Update  : 2025/4/26 3:05
+# @Detail  : 
 
-from .Binary.reader import BinaryReader
+from .binary import BinaryReader
 
 __all__ = [
     'BinaryReader'
