+++ src/league_tools/formats/__init__.py
@@ -4,13 +4,13 @@
 # @Site    : x-item.com
 # @Software: PyCharm
 # @Create  : 2021/3/4 18:46
-# @Update  : 2024/9/2 8:25
+# @Update  : 2025/4/26 3:09
 # @Detail  : 
 
-from .BIN import BIN, StringHash
-from .BNK import BNK, HIRC
-from .WAD import WAD, WadHeaderAnalyzer
-from .WPK import WPK
+from src.league_tools.formats.bin.parser import BIN, StringHash
+from src.league_tools.formats.bnk.parser import BNK, HIRC
+from src.league_tools.formats.wad.parser import WAD, WadHeaderAnalyzer
+from src.league_tools.formats.wpk.parser import WPK
 
 __all__ = [
     'BIN',
