+++ src/league_tools/formats/section/DIDX.py
@@ -4,11 +4,11 @@
 # @Site    : x-item.com
 # @Software: PyCharm
 # @Create  : 2021/3/1 21:09
-# @Update  : 2025/2/9 4:15
+# @Update  : 2025/4/26 3:06
 # @Detail  : Wwise bnk文件, DIDX块
 
 
-from src.league_tools.base import SectionNoIdBNK, WemFile
+from src.league_tools.core.section import SectionNoIdBNK, WemFile
 
 
 class DIDX(SectionNoIdBNK):
