+++ src/league_tools/formats/wpk/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+# @Author  : Virace
+# @Email   : Virace@aliyun.com
+# @Site    : x-item.com
+# @Software: Pycharm
+# @Create  : 2025/4/26 3:08
+# @Update  : 2025/4/26 3:08
+# @Detail  :
